+++ backend/src/features/challenges/structs/requests/challenge_daily_tracking.rs
@@ -25,7 +25,7 @@ pub struct GetMultipleChallengesDailyTrackingsRequest {
 pub struct ChallengeDailyTrackingUpdateRequest {
     pub day_of_program: i32,
     pub habit_id: Uuid,
-    pub quantity_per_set: i32,
+    pub quantity_per_set: f64,
     pub quantity_of_set: i32,
     pub unit_id: Uuid,
     pub weight: i32,
@@ -38,7 +38,7 @@ pub struct ChallengeDailyTrackingCreateRequest {
     pub challenge_id: Uuid,
     pub habit_id: Uuid,
     pub day_of_program: i32,
-    pub quantity_per_set: i32,
+    pub quantity_per_set: f64,
     pub quantity_of_set: i32,
     pub unit_id: Uuid,
     pub weight: i32,
