+++ frontend/lib/features/habits/data/repositories/habit_daily_tracking_repository_impl.dart
@@ -62,7 +62,7 @@ class HabitDailyTrackingRepositoryImpl implements HabitDailyTrackingRepository {
   Future<Either<DomainError, HabitDailyTracking>> createHabitDailyTracking({
     required String habitId,
     required DateTime datetime,
-    required int quantityPerSet,
+    required double quantityPerSet,
     required int quantityOfSet,
     required String unitId,
     required int weight,
@@ -112,7 +112,7 @@ class HabitDailyTrackingRepositoryImpl implements HabitDailyTrackingRepository {
   Future<Either<DomainError, HabitDailyTracking>> updateHabitDailyTracking({
     required String habitDailyTrackingId,
     required DateTime datetime,
-    required int quantityPerSet,
+    required double quantityPerSet,
     required int quantityOfSet,
     required String unitId,
     required int weight,
